[backend] init jwt

+++ backend/pkg/handlers/user.go
@@ -3,7 +3,6 @@ package handlers
 import (
 	"gossip/backend/pkg/models"
 	"gossip/backend/pkg/utils"
-	"net/http"
 
 	"github.com/gofiber/fiber/v2"
 	"gorm.io/gorm"
@@ -17,13 +16,19 @@ func (h *UserHandler) GetUserById(c *fiber.Ctx) error {
 	var err error
 	var user models.User
 	id := c.Params("id")
+	currId := c.Get("user_id")
 
 	// get user by id
 	if err = h.DB.Where("id = ?", id).First(&user).Error; err != nil {
-		return fiber.NewError(http.StatusNotFound, "User not found")
+		return fiber.NewError(fiber.StatusNotFound, "User not found")
 	}
 
-	return c.Status(http.StatusOK).JSON(fiber.Map{"data": user})
+	// check authorised
+	if !user.IsPublic && currId != user.ID {
+		return fiber.NewError(fiber.StatusUnauthorized, "Unauthorised")
+	}
+
+	return c.Status(fiber.StatusOK).JSON(fiber.Map{"data": user})
 }
 
 func (h *UserHandler) UpdateUser(c *fiber.Ctx) error {
@@ -36,14 +41,20 @@ func (h *UserHandler) UpdateUser(c *fiber.Ctx) error {
 	var input updateUserInput
 	var user models.User
 	id := c.Params("id")
+	currId := c.Get("user_id")
 
 	// get user by id
 	if err = h.DB.Where("id = ?", id).First(&user).Error; err != nil {
-		return fiber.NewError(http.StatusNotFound, "User not found")
+		return fiber.NewError(fiber.StatusNotFound, "User not found")
+	}
+
+	// check authorised
+	if currId != user.ID {
+		return fiber.NewError(fiber.StatusUnauthorized, "Unauthorised")
 	}
 
 	if err = c.BodyParser(&input); err != nil {
-		return fiber.NewError(http.StatusBadRequest, "Invalid fields")
+		return fiber.NewError(fiber.StatusBadRequest, "Invalid fields")
 	}
 
 	// input validation
@@ -54,7 +65,7 @@ func (h *UserHandler) UpdateUser(c *fiber.Ctx) error {
 	// only hash new password if provided
 	hashedPassword, err := utils.HashPassword(input.Password)
 	if err != nil {
-		return fiber.NewError(http.StatusInternalServerError, err.Error())
+		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
 	}
 
 	updateUser := models.User{
@@ -64,28 +75,34 @@ func (h *UserHandler) UpdateUser(c *fiber.Ctx) error {
 
 	// update user
 	if err = h.DB.Model(&user).Updates(updateUser).Error; err != nil {
-		return fiber.NewError(http.StatusInternalServerError, err.Error())
+		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
 	}
 
-	return c.Status(http.StatusOK).JSON(fiber.Map{"data": user})
+	return c.Status(fiber.StatusOK).JSON(fiber.Map{"data": user})
 }
 
 func (h *UserHandler) DeleteUser(c *fiber.Ctx) error {
 	var err error
 	var user models.User
 	id := c.Params("id")
+	currId := c.Get("user_id")
 
 	// get user by id
 	if err = h.DB.Where("id = ?", id).First(&user).Error; err != nil {
-		return fiber.NewError(http.StatusNotFound, "User not found")
+		return fiber.NewError(fiber.StatusNotFound, "User not found")
+	}
+
+	// check authorised
+	if currId != user.ID {
+		return fiber.NewError(fiber.StatusUnauthorized, "Unauthorised")
 	}
 
 	// delete user
 	if err = h.DB.Delete(&user).Error; err != nil {
-		return fiber.NewError(http.StatusInternalServerError, err.Error())
+		return fiber.NewError(fiber.StatusInternalServerError, err.Error())
 	}
 
-	return c.Status(http.StatusOK).JSON(fiber.Map{"data": true})
+	return c.Status(fiber.StatusOK).JSON(fiber.Map{"data": true})
 }
 
 func (h *UserHandler) ToggleProfileVisibility(c *fiber.Ctx) error {
